deploy using common Handler
